Slashcommands & Moderation Features

+++ src/commands/purge.ts
@@ -4,37 +4,41 @@ import Command from "../struct/Command"
 import Roles from "../util/roles"
 import hexToRGB from "../util/hexToRGB"
 import Discord from "discord.js"
+import CommandMessage from "../struct/CommandMessage"
 
 export default new Command({
     name: "purge",
     aliases: ["prune", "bulkdelete"],
     description: "Bulk-delete messages in a channel.",
     permission: [Roles.MODERATOR, Roles.MANAGER],
-    usage: "<amount>",
-    async run(this: Command, client: Client, message: Discord.Message, args: Args) {
-        const amount = Number(args.consume())
+    args: [
+        {
+            name: "amount",
+            description: "Amount to purge.",
+            required: true,
+            optionType: "INTEGER"
+        }
+    ],
+    async run(this: Command, client: Client, message: CommandMessage, args: Args) {
+        const amount = Number(args.consume("amount"))
+        if (!amount)
+            return client.response.sendError(message, client.messages.invalidAmount)
         if (Number.isNaN(amount))
-            return client.channel.sendError(
-                message.channel,
-                "You must provide a valid amount!"
-            )
+            return client.response.sendError(message, client.messages.invalidAmount)
         if (amount > 100)
-            return client.channel.sendError(
-                message.channel,
-                "You can only purge 100 messages at a time!"
-            )
+            return client.response.sendError(message, client.messages.purgeLimit)
+
+        await message.continue()
+
         const purged = await (message.channel as Discord.TextChannel).bulkDelete(
             amount,
             true
         )
-        await client.channel.sendSuccess(
-            message.channel,
-            `Purged ${purged.size} messages.`
-        )
+        await client.response.sendSuccess(message, `Purged ${purged.size} messages.`)
         await client.log({
             color: hexToRGB(client.config.colors.info),
             author: { name: "Purge" },
-            description: `${purged.size} messages purged by ${message.author} in ${message.channel}.`
+            description: `${purged.size} messages purged by ${message.member} in ${message.channel}.`
         })
     }
 })
